Further optimized node selector panel
